wouhou c'est tout bo

+++ src/pt/alextd1/model/Message.java
@@ -1,5 +1,7 @@
 package pt.alextd1.model;
 
+import org.germain.tool.ManaBox;
+import pt.alextd1.tools.KeyGen;
 import pt.alextd1.tools.Transcoder;
 
 import java.io.IOException;
@@ -12,23 +14,18 @@ public class Message
 {
     // déclaration des membres utilisés pour encoder et décoder des messages
     // ici le boolean qui déterminera si le message est codé ou non, et donc s'il faut le coder ou le décoder
-    private boolean encoded;
+    private final boolean encoded;
 
-    // liste des lignes du message non codé - inutilisé car récupéré directement
-    private List<String> msgClear;
-
-    // pareil qu'au dessus pour le message décodé
-    private List<String> msgEncodded;
 
     // chemin du fichier copntenant le message décodé
-    private Path msgClearPath;
+    private final Path msgClearPath;
 
 
     // chemin du fichier copntenant le message encodé
-    private Path msgEncodedPath;
+    private final Path msgEncodedPath;
 
     // chemin de la clef de décryptage
-    private Path keyPath;
+    private final Path keyPath;
 
 
     // clef de décryptage
@@ -52,9 +49,9 @@ public class Message
     public void editer()
     {
         // si encoded est a true, le message est donc codé, donc on le décode
-        if(encoded){
+        if (encoded) {
             try {
-                // on déclare une nouvelle liste qui sotckera toutes les lignes du fichier contenant le message codé
+                // on déclare une nouvelle liste qui stockera toutes les lignes du fichier contenant le message codé
                 List<String> lireCode = Files.readAllLines(msgEncodedPath);
                 //on récupère la clef
                 key = Files.readString(keyPath);
@@ -62,16 +59,16 @@ public class Message
                 transCoder = new Transcoder(key);
 
                 // pour chaque chaine de caractère de la liste, chaque ligne de message codé
-                for(String chaine : lireCode){
+                for (String chaine : lireCode) {
 
                     // écrire dans le fichier de décodage le message décodé. s'ajoute ligne par ligne grace au  APPEND.
-                    Files.writeString(msgClearPath, transCoder.decode(chaine)+System.lineSeparator(),StandardOpenOption.CREATE,StandardOpenOption.APPEND);
+                    Files.writeString(msgClearPath, transCoder.decode(chaine) + System.lineSeparator(), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
                 }
 
             }
             catch (IOException e) {
                 // récupère les erreurs et affiche les différentes données pour la localiser si besoin
-                System.out.println("erreur " + e.getMessage()+" \n "+e.getCause());
+                System.out.println("erreur " + e.getMessage() + " \n " + e.getCause());
                 System.out.println(msgClearPath);
                 System.out.println(key);
                 System.out.println(keyPath);
@@ -79,37 +76,66 @@ public class Message
             }
 
 
-        }else{
+        } else {
             // sinon encodé est false : donc le message n'est pas codé et il faut l'encoder
+            // déclaration d'une liste contenant toutes les lignes de message non codé
+            List<String> lireDecode = null;
             try {
-                // déclaration d'une liste contenant toutes les lignes de message non codé
-                List<String> lireDecode = Files.readAllLines(msgClearPath);
+                lireDecode = Files.readAllLines(msgClearPath);
+            }
+            catch (IOException e) {
+                e.printStackTrace();
+            }
 
+            if (Files.exists(keyPath)) {
                 // récupération de la clef, comme au dessus
-                key = Files.readString(keyPath);
+                try {
+                    key = Files.readString(keyPath);
+                }
+                catch (IOException e) {
+                    e.printStackTrace();
+                }
                 // instanciation du transcoder
                 transCoder = new Transcoder(key);
 
                 // pour chaque ligne de message non codé
-                for(String chaine : lireDecode){
+                for (String chaine : lireDecode) {
                     //écrire la version codé dans un fichier
-                    Files.writeString(msgEncodedPath, transCoder.encode(chaine)+System.lineSeparator(),StandardOpenOption.CREATE,StandardOpenOption.APPEND);
+                    try {
+                        Files.writeString(msgEncodedPath, transCoder.encode(chaine) + System.lineSeparator(), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+                    }
+                    catch (IOException e) {
+                        e.printStackTrace();
+                    }
                 }
+            } else {
+                KeyGen genclef = new KeyGen();
+                try {
+                    Files.writeString(keyPath, ManaBox.encrypt(genclef.randomKey()), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+                }
+                catch (IOException e) {
+                    e.printStackTrace();
+                }
+                try {
+                    key = Files.readString(keyPath);
+                }
+                catch (IOException e) {
+                    e.printStackTrace();
+                }
+                // instanciation du transcoder
+                transCoder = new Transcoder(key);
 
+                // pour chaque ligne de message non codé
+                for (String chaine : lireDecode) {
+                    //écrire la version codé dans un fichier
+                    try {
+                        Files.writeString(msgEncodedPath, transCoder.encode(chaine) + System.lineSeparator(), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+                    }
+                    catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
             }
-            catch (IOException e) {
-                // erreurs + données, cf au dessus
-                System.out.println("erreur" + e.getMessage()+"\n"+e.getCause());
-                System.out.println(msgClearPath);
-                System.out.println(key);
-                System.out.println(keyPath);
-                System.out.println(msgEncodedPath);
-            }
-
         }
     }
-
-
-
-
 }
